Moved Reek Examiner encapsulation into own modules
